gson model deserialization cleanup
