+++ aws-api/src/main/java/com/amplifyframework/api/aws/GsonFactory.java
@@ -55,6 +55,8 @@ final class GsonFactory {
         ModelWithMetadataAdapter.register(builder);
         SerializedModelAdapter.register(builder);
         SerializedCustomTypeAdapter.register(builder);
+        ModelListDeserializer.register(builder);
+        ModelPageDeserializer.register(builder);
         builder.serializeNulls();
         return builder.create();
     }
